+++ catalog/app/components/Preview/loaders/Ngl.tsx
@@ -1,3 +1,5 @@
+import { extname } from 'path'
+
 import type { PromiseResult } from 'aws-sdk/lib/request'
 import * as R from 'ramda'
 import * as React from 'react'
@@ -9,7 +11,10 @@ import { PreviewData } from '../types'
 
 import * as utils from './utils'
 
-export const detect = R.pipe(utils.stripCompression, utils.extIn(['.ent', '.pdb']))
+export const detect = R.pipe(
+  utils.stripCompression,
+  utils.extIn(['.cif', '.ent', '.mol', '.mol2', '.pdb', '.sdf']),
+)
 
 const gzipDecompress = DecompressorRegistry.get('gz')
 
@@ -25,7 +30,8 @@ export const Loader = function NglLoader({ handle, children }: NglLoaderProps) {
     (r: PromiseResult<{ Body: Uint8Array | string }, null>) => {
       const compression = utils.getCompression(handle.key)
       const body = compression === 'gz' ? gzipDecompress(r.Body as string) : r.Body
-      return PreviewData.Ngl({ blob: new Blob([body]) })
+      const ext = extname(utils.stripCompression(handle.key)).substring(1)
+      return PreviewData.Ngl({ blob: new Blob([body]), ext })
     },
   )
   const handled = utils.useErrorHandling(processed, { handle, retry: data.fetch })