+++ catalog/app/containers/Admin/UsersAndRoles.js
@@ -5,7 +5,7 @@ import * as APIConnector from 'utils/APIConnector'
 import MetaTitle from 'utils/MetaTitle'
 import * as Cache from 'utils/ResourceCache'
 
-import Roles from './Roles'
+import { Roles, Policies } from './RolesAndPolicies'
 import Users from './Users'
 import * as data from './data'
 
@@ -15,12 +15,15 @@ export default function UsersAndRoles() {
   const users = Cache.useData(data.UsersResource, { req })
   return (
     <>
-      <MetaTitle>{['Users and Roles', 'Admin']}</MetaTitle>
+      <MetaTitle>{['Users, Roles and Policies', 'Admin']}</MetaTitle>
       <M.Box mt={2}>
         <Users users={users} />
       </M.Box>
       <M.Box mt={2} mb={2}>
         <Roles />
+      </M.Box>
+      <M.Box mt={2} mb={2}>
+        <Policies />
       </M.Box>
     </>
   )