+++ catalog/app/containers/Bucket/PackageDialog/PackageDialog.tsx
@@ -1,6 +1,5 @@
 import { basename } from 'path'
 
-import { FORM_ERROR } from 'final-form'
 import * as R from 'ramda'
 import * as React from 'react'
 import type * as RF from 'react-final-form'
@@ -13,6 +12,7 @@ import { useData } from 'utils/Data'
 import * as APIConnector from 'utils/APIConnector'
 import * as AWS from 'utils/AWS'
 import * as Sentry from 'utils/Sentry'
+import { mkFormError } from 'utils/formTools'
 import { JsonSchema, makeSchemaValidator } from 'utils/json-schema'
 import * as packageHandleUtils from 'utils/packageHandle'
 import * as s3paths from 'utils/s3paths'
@@ -34,25 +34,6 @@ export const ERROR_MESSAGES = {
 export const getNormalizedPath = (f: { path?: string; name: string }) => {
   const p = f.path || f.name
   return p.startsWith('/') ? p.substring(1) : p
-}
-
-export const mkFormError = (err: React.ReactNode) => ({ [FORM_ERROR]: err })
-
-interface InputError {
-  path: string | null
-  message: string
-}
-
-export function mapInputErrors(
-  inputErrors: Readonly<InputError[]>,
-  mapping: Record<string, string> = {},
-) {
-  const formErrors: Record<string, string> = {}
-  for (let err of inputErrors) {
-    const key = err.path && err.path in mapping ? mapping[err.path] : FORM_ERROR
-    formErrors[key] = err.message
-  }
-  return formErrors
 }
 
 export async function hashFile(file: File) {